+++ packages/core/lib/core.dart
@@ -14,3 +14,4 @@ export 'src/repositories/budget_repository.dart';
 export 'src/usecases/add_transaction.dart';
 export 'src/usecases/get_balance.dart';
 export 'src/usecases/import_transactions.dart';
+export 'src/utils/budget_period_calculator.dart';